refactor(cmd): move prompt to util package
